ethcheck
This code can be used to quickly retrieve OSNIT from the Ethereum blockchain from your terminal without having to go to a website yourself. Manually searching for account on etherscan can be time consuming; there are many buttons to press in the browser. So I wanted to just do it from the terminal with one command. The code as seen here would retrieve information about how much money is in a particular know Ethereum public address (the address would need to be inputed as an argument and has to be in the correct format, with a '0x' in the front indicating hexadecimal format). This code is written in python and essentially functions as a web scraper. It could obviously be modified to retrieve any information about a particular account, but I just chose to print the amount of money. After a user inputs the address and runs the script, that argument would be appended to the end of the URL variable. Then a curl command is run as a subprocess with that URL and its HTML is saved into the gettext variable. For identifying the amount of money in the account, I had to check for patterns in the HTML. Luckily all funds came after the string 8">$ which is the only string of that type in all accounts' HTML. Once the for loop found that string it saved the content thereafter, until the '<', into the fund variable. Then I just simply printed the funds. This script can be a quick way to check the funds of any Ethereum account and takes about 2 seconds to run. It can be handy if one has a bunch of accounts they need to investigate.